Move a bunch of stuff into a new `guts` folder for parts which aren't intended to be used by end users.
