+++ selfie-lib/src/commonMain/kotlin/com/diffplug/selfie/guts/Slice.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019-2023 DiffPlug
+ * Copyright (C) 2019-2024 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.diffplug.selfie
+package com.diffplug.selfie.guts
 
 internal class Slice(val base: String, val startIndex: Int = 0, val endIndex: Int = base.length) :
     CharSequence {
