+++ selfie-lib/src/commonMain/kotlin/com/diffplug/selfie/guts/SnapshotStorage.kt
@@ -13,7 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.diffplug.selfie
+package com.diffplug.selfie.guts
+
+import com.diffplug.selfie.ExpectedActual
+import com.diffplug.selfie.Snapshot
 
 expect interface Path : Comparable<Path>
