+++ selfie-lib/src/commonMain/kotlin/com/diffplug/selfie/guts/WriteTracker.kt
@@ -13,7 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.diffplug.selfie
+package com.diffplug.selfie.guts
+
+import com.diffplug.selfie.Snapshot
 
 expect class CallLocation : Comparable<CallLocation> {
   val file: String?
