--- selfie-lib/src/commonTest/kotlin/com/diffplug/selfie/SliceTest.kt
@@ -1,33 +1,0 @@
-/*
- * Copyright (C) 2020-2023 DiffPlug
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.diffplug.selfie
-
-import io.kotest.matchers.shouldBe
-import kotlin.test.Test
-
-class SliceTest {
-  @Test
-  fun unixLine() {
-    Slice("A single line").unixLine(1).toString() shouldBe "A single line"
-    val oneTwoThree = Slice("\nI am the first\nI, the second\n\nFOURTH\n")
-    oneTwoThree.unixLine(1).toString() shouldBe ""
-    oneTwoThree.unixLine(2).toString() shouldBe "I am the first"
-    oneTwoThree.unixLine(3).toString() shouldBe "I, the second"
-    oneTwoThree.unixLine(4).toString() shouldBe ""
-    oneTwoThree.unixLine(5).toString() shouldBe "FOURTH"
-    oneTwoThree.unixLine(6).toString() shouldBe ""
-  }
-}
