+++ selfie-lib/src/jsMain/kotlin/com/diffplug/selfie/guts/SnapshotStorage.js.kt
@@ -13,15 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.diffplug.selfie
+package com.diffplug.selfie.guts
 actual fun initStorage(): SnapshotStorage {
-  try {
-    val clazz = Class.forName("com.diffplug.selfie.junit5.SnapshotStorageJUnit5")
-    return clazz.getMethod("initStorage").invoke(null) as SnapshotStorage
-  } catch (e: ClassNotFoundException) {
-    throw IllegalStateException(
-        "Missing required artifact `com.diffplug.spotless:selfie-runner-junit5", e)
-  }
+  TODO("Not yet implemented")
 }
 
-actual typealias Path = java.nio.file.Path
+actual typealias Path = CharSequence
